Run doctest in all temporian/*.py and docs/*.md
